==> Code/2- Clean 2019 data.ipynb <==
# Importing packages
import pandas as pd
pd.options.mode.chained_assignment = None
import numpy as np
from datetime import datetime, timedelta
from sqlalchemy import create_engine
import matplotlib.pyplot as plt
import seaborn as sns 

# Accessing 2019 data from database
engine = create_engine('sqlite:///mta_2019.db')
df = pd.read_sql('SELECT * FROM mta_2019;',engine)


df.head()


#set first line as header

new_header = df.iloc[0] 
df = df[1:] 
df.columns = new_header 

#reset index
df.reset_index(drop=True)


# Rename columns to remove whitespace
df.columns = df.columns.str.replace(" ", "")

# Create DATETIME column from 'DATE' and 'TIME' columns
df['DATETIME'] = (df['DATE'] + " " + df['TIME'])
df["DATETIME"] = pd.to_datetime(df["DATETIME"], errors='coerce', format='%m/%d/%Y %H:%M:%S')
df.head()

#get date per row 
df["DATE"] = df["DATETIME"].dt.date

#get week per row
df["WEEK"] = df["DATETIME"].dt.isocalendar().week

df.head()

df.dtypes

#convert entries and exits to int64
df["ENTRIES"] = pd.to_numeric(df["ENTRIES"], errors='coerce').convert_dtypes() 
df["EXITS"] = pd.to_numeric(df["EXITS"], errors='coerce').convert_dtypes() 
df.dtypes

df.DESC.unique()

# Remove non-REGULAR values from 'DESC'
df = df.drop(df.loc[df.DESC != 'REGULAR'].index)
df.DESC.unique()

df.head()

#clean station names 
df["STATION"] = df["STATION"].str.replace(' ', '_')
df["STATION"] = df["STATION"].str.replace('-', '_')

df.head()

#identify duplicate stations
lines = df.groupby(["STATION","LINENAME"]).first().reset_index()
lines.head()

#create station + lines columns
df["STATION_FULL"] = df["STATION"] + "_Line_" + df["LINENAME"]
df.head()

# Create previous datetime, entries and exits columns

df[["PREV_DATETIME", "PREV_ENTRIES", "PREV_EXITS"]] = (df
                                                .groupby(['STATION_FULL'])
                                                       ['DATETIME','ENTRIES','EXITS']
                                                .apply(lambda grp: grp.shift(1)))

df.head()

# Drop Na rows
print("Before drop NA: ", df.shape)
df.dropna(how='any', inplace=True)
print("After drop NA: ", df.shape)

#count the actual number of entries and exits from cumulative numbers 
df["ENTRY_COUNT"] = abs(df["ENTRIES"] - df["PREV_ENTRIES"])
df["EXIT_COUNT"] = abs(df["EXITS"] - df["PREV_EXITS"])

#clean entry count for values in excess of 14400, equal to one person per sec in 4 hour interval

index_entry = df[ df['ENTRY_COUNT'] > 14400 ].index
df.drop(index_entry , inplace=True)

#clean exit count for values in excess of 14400, equal to one person per sec in 4 hours interval 

index_exit = df[df['EXIT_COUNT'] > 14400].index
df.drop(index_exit, inplace = True)

#sanity check
df.EXIT_COUNT.describe()

#sanity check
df.ENTRY_COUNT.describe()

#get traffic per row
df["TRAFFIC"] = df["ENTRY_COUNT"] + df["EXIT_COUNT"]


df.head()

#sanity check
df.TRAFFIC.describe()

# check for nan
print("Before drop NA: ", df.shape)
df.dropna(how='any', inplace=True)
print("After drop NA: ", df.shape)

#identify duplicate stations
df.groupby(["STATION_FULL"]).first()

#get traffic per date per station
df_daily = df.groupby(["STATION_FULL", "DATE"]).sum("TRAFFIC").reset_index()
df_daily.head()

# Drop unnecessary columns from daily df 
df_daily = df_daily.drop(["ENTRIES", "EXITS", "PREV_ENTRIES", "PREV_EXITS", "ENTRY_COUNT", "EXIT_COUNT","WEEK"], axis=1)
df_daily.head()

#make sure the last week is complete 
df_daily.DATE.tail(10)
#week 52 is december 23-29,2019 therefore week 52 data is incomplete 
#week 52 will be dropped from analysis for weekly traffic 

#make sure week 1 is complete 
df_daily.DATE.head(10)
#week 1 is 12/31/2018 - 1/6/2019 therefore 
#week 1 data is incomplete and will be dropped from analysis for
#weekly traffic

#get traffic per week per station
df_weekly = df.groupby(["STATION_FULL", "WEEK"]).sum("TRAFFIC").reset_index()
df_weekly.head()

#remove week 1 data, as it is incomplete
df_weekly = df_weekly.drop(df_weekly.loc[df_weekly.WEEK == 1 ].index)
df_weekly.WEEK.unique()

#remove week 31-52 data, as 52 is incomplete and we only use up to week 30 in 2021 data
no_weeks=list(range(31,53))
df_weekly = df_weekly.drop(df_weekly.loc[df_weekly.WEEK > 30 ].index)
df_weekly.WEEK.unique()


# Drop unnecessary columns from weekly df 
df_weekly = df_weekly.drop(["ENTRIES", "EXITS", "PREV_ENTRIES", "PREV_EXITS", "ENTRY_COUNT", "EXIT_COUNT"], axis=1)
df_weekly.head()

df_weekly.shape


df_weekly.dtypes

#save data for 2019 for further use 
df_weekly.to_csv("./all_traffic_2019.csv", index=False)

#get average traffic over all stations for 2019

df_2019 = df_weekly.groupby("WEEK").mean("TRAFFIC").reset_index()
df_2019["STATION"] = "AVERAGE"
df_2019.head()

#load top ten from 2021 
topten_2021 = pd.read_csv("./top_traffic_2021.csv")
topten_2021.head()


#get list of top_ten_2021
tt_2021_list=topten_2021["STATION_FULL"].unique()
tt_2021_list

#create filter for top stations of 2021
top2021_filter = df_weekly["STATION_FULL"].isin(tt_2021_list)
top2021_filter

#get df of traffic data per week for top stations 
df_top_traffic_2021 = df_weekly[top2021_filter]
df_top_traffic_2021.head()

#save df for further use 
df_top_traffic_2021.to_csv("./top_2021_in_2019.csv", index=False)
