==> Code/3- Calculate percentage.ipynb <==
import pandas as pd
import urllib.request
import os
import sqlite3
from pathlib import Path
import requests
import json
import time

#import top ten data set with full line names 
tt_2021 = pd.read_csv("./top_traffic_2021.csv")
tt_2021.head()

#load traffic data from 2019 for top ten in 2021 
tt_2019 = pd.read_csv("./top_2021_in_2019.csv")
tt_2019.head()

station_list = tt_2019.STATION_FULL.unique()
station_list

#create individual dataframe per station that combines 2019 data and 2021 data in order to calculate percentage
#of traffic per week per station

for station in station_list:
        
        filter_2019 = ()
        filter_2021 = ()
        column_names = ["STATION_FULL","WEEK","TRAFFIC"]
        data_2019 = pd.DataFrame(columns = column_names)
        data_2021 = pd.DataFrame(columns = column_names)
        
#Create a unique DF per top station for its data in 2019

        filter_2019 = tt_2019["STATION_FULL"]== station
        data_2019 = tt_2019[filter_2019].reset_index()
        data_2019["STATION_FULL"] = data_2019["STATION_FULL"] + "_2019"

#Create a unique df per top station for its data in 2021
        filter_2021 = tt_2021["STATION_FULL"]== station
        data_2021 = tt_2021[filter_2021]
        data_2021 = data_2021[data_2021.WEEK != 1]
        data_2021["STATION_FULL"] = data_2021["STATION_FULL"] + "_2021"
          
#create one df for both years 

        data_all= pd.concat([data_2019,data_2021],ignore_index=True)
    
#create percentage df for 2021
        
        percent = data_all.groupby(["WEEK","STATION_FULL"])["TRAFFIC"].first().reset_index()
        percent["PERCENT"] = (percent["TRAFFIC"].div(percent.TRAFFIC.shift(1)))*100
        percent['PERCENT'].values[percent['PERCENT'].values > 100] = 0
        percent['PERCENT'] = percent['PERCENT'].fillna(0)
        percent = percent[percent["PERCENT"] > 0]
        percent.to_csv("./percent/Percent_"+ station+"_2021.csv", index = False)

#save traffic data from 2021 and 2019 for further use 

        data_all.to_csv("./station_data/"+station+"_2021_2019.csv", index=False)

#combine all 2021 and 2019 traffic data df's into one csv

dirpath = (os.path.expanduser("./station_data/"))
output = (os.path.expanduser("./top_traffic_2021_in_2019.csv"))
csvout_lst = []
files = [os.path.join(dirpath, fname) for fname in os.listdir(dirpath)]

for filename in sorted(files):
    data = pd.read_csv(filename, sep=',', index_col=0, header=0)
    csvout_lst.append(data)

pd.concat(csvout_lst).to_csv(output)

#combine all percent df's into one csv

dirpath = (os.path.expanduser("./percent/"))
output = (os.path.expanduser("./top_percent_2021_in_2019.csv"))
csvout_lst = []
files = [os.path.join(dirpath, fname) for fname in os.listdir(dirpath)]

for filename in sorted(files):
    data = pd.read_csv(filename, sep=',', index_col=0, header=0)
    csvout_lst.append(data)

pd.concat(csvout_lst).to_csv(output)
