==> Code/5- Identify Starbucks locations .ipynb <==
import pandas as pd
import urllib.request
import os
import sqlite3
from pathlib import Path
import requests
import json
import time
import numpy as np
from scipy import spatial

#starting with importing the data for our top 5 stations 
#import top ten df
top_five = pd.read_csv('top_5_stations.csv')
top_five.head()

top_five = top_five[top_five["WEEK"]==30].sort_values("TRAFFIC", ascending = False).reset_index(drop=True)
top_five.head()


#add rank column according to traffic 

top_five["T_RANK"] = top_five.index+1
top_five.head()

#create a new data frame of station names and remove the "_2021" designation 

stations = top_five.STATION_FULL.str[:-5].unique()
top_five_loc = pd.DataFrame({"STATION": stations})
top_five_loc['STATION'] = top_five_loc['STATION'].str.split('_Line').str[0]

#add rank column according to traffic


top_five_loc.head()

#change names of stations for google api match

top_five_loc.loc[0, 'STATION'] = "GRAND_CENTRAL"
top_five_loc.loc[2, 'STATION'] = "42_ST_PORT_AUTHORITY"
top_five_loc.loc[3, 'STATION'] = "PENN_STATION_ACE"
top_five_loc.head()


#create places variable ready for passing to url 

API_KEY = 
places = top_five_loc["STATION"].str.replace("_", "+") + "+mta+subway+NY+NY"
req1 = "https://maps.googleapis.com/maps/api/geocode/json?address="
req2 = "&key=" + API_KEY

#function for getting coordinates for the stations from google api 

def get_loc(places,req1,req2):

    locs = {}
    counter = 0
    
    for place in places:
        r = requests.get(req1 + place + req2)
        
        try:
            results = r.json()['results'][0][u'geometry'][u'location']
            lat = results[u'lat']
            lng = results[u'lng']
            locs[place] = [lat,lng]
            
        except:
            pass
            
    return locs

#get coordinates for top 5 stations

loc = get_loc(places,req1,req2)

#convert dictionary to dataframe and transpose 



loc_five = pd.DataFrame(loc)
loc_five = loc_five.T.reset_index()

#rename columns

loc_five = loc_five.rename(columns={"index": "STATION", 0: "LAT", 1:"LONG"})

#change names back to match traffic data 

loc_five["STATION"] = top_five.STATION_FULL.unique()


loc_five.head()

#function for getting starbucks locations for stations of interest

def get_bucks(df):

    for i,rows in df.iterrows():
    
     
        
            API_KEY = 
            req1 = "https://maps.googleapis.com/maps/api/place/textsearch/json?query=starbucks&location="
            req2 = "&radius=10&region=us&type=cafe&key=" + API_KEY
            endpoint= req1 + str(df.loc[i,'LAT'])+","+str(df.loc[i,'LONG']) + req2
            station = df.loc[i,"STATION"]
            
        
            coffee_shops = []
            params = {}
  
            endpoint_url = endpoint
            res = requests.get(endpoint_url, params = params)
            results =  json.loads(res.content)
            coffee_shops.extend(results['results'])
            time.sleep(2)
            
            while "next_page_token" in results:
                    params['pagetoken'] = results['next_page_token'],
                    res = requests.get(endpoint_url, params = params)
                    results = json.loads(res.content)
                    coffee_shops.extend(results['results'])
                    time.sleep(2)
        
            shop_name = []
            shop_address = []
            shop_lat = []
            shop_lon = []
            business_status = []


            for i in range(len(coffee_shops)): 
    
                shop = coffee_shops[i]
    
                try:
                        shop_name.append(shop['name'])
                except:
                        shop_name.append('none')
        
                try: 
                        shop_address.append(shop['formatted_address'])
                except:
                        shop_address.append('none')
        
    
                try: 
                        shop_lat.append(shop['geometry']['location']['lat'])
                except:
                        shop_lat.append('none')
        
                try: 
                        shop_lon.append(shop['geometry']['location']['lng'])
                except:
                        shop_lon.append('none')
        
                try:
                        business_status.append(shop['business_status'])
                    
                except:
                        business_status.append('none')
                    
                    
            df_dict = {'NAME': shop_name, "ADDRESS":shop_address, "LAT" : shop_lat, "LONG": shop_lon, "STATUS": business_status} 
            output = pd.DataFrame(df_dict)
            
            
            
            return(output)
            

#get starbucks locations for station of interest using google api 
#WTC

def get_nearest(station):



    starbucks_station = get_bucks(station)


#clean the data set & make sure there are no duplicates

    starbucks = starbucks_station.copy()
    starbucks["ADDRESS"] = starbucks["ADDRESS"].str[:-5]
    starbucks = starbucks.drop_duplicates(subset=['ADDRESS']).reset_index(drop=True)
    starbucks.ADDRESS.nunique()

#find the closest three starbucks shops for our location of interest
#create tuple and index from starbuckscoordinate data for passing to kdtree 


    starbucks_coords = list(zip(starbucks.LAT, starbucks.LONG))
    starbucks_index = starbucks.index

#create the reference coordinates for the KDtree seach using WTC data from our location df 
    loc_coords = list(zip(station.LAT, station.LONG))
    

#Run KDTree search on our list of Starbucks locations and identify 
#top 3 closest locations 


    tree = spatial.KDTree(starbucks_coords)
    idxs  = tree.query(loc_coords, 3)
    locations = [starbucks_index[idx] for idx in idxs[1]]
    
    #create df of results from KDTree
    shop_ind = locations[0].tolist()
    shop_locations = starbucks.iloc[shop_ind].reset_index(drop=True)
    
    #add station name to df 
    shop_locations["STATION"] = station.loc[0,"STATION"]
    
    
    
    return(shop_locations)



#Starbucks locations for station 1 in top 5 list 

station = pd.DataFrame(loc_five.loc[0,:]).T.reset_index()
starbucks_1 = get_nearest(station)
starbucks_1

#Starbucks locations for station 2 in top 5 list 

station = pd.DataFrame(loc_five.loc[1,:]).T.reset_index()
starbucks_2 = get_nearest(station)
starbucks_2

#Starbucks locations for station 3 in top 5 list 

station = pd.DataFrame(loc_five.loc[2,:]).T.reset_index()
starbucks_3 = get_nearest(station)
starbucks_3

#Starbucks locations for station 4 in top 5 list 

station = pd.DataFrame(loc_five.loc[3,:]).T.reset_index()
starbucks_4 = get_nearest(station)
starbucks_4

#Starbucks locations for station 5 in top 5 list 

station = pd.DataFrame(loc_five.loc[4,:]).T.reset_index()
starbucks_5 = get_nearest(station)
starbucks_5

#concat all starbucks locations into one dataframe

starbucks_ind = [starbucks_1, starbucks_2, starbucks_3, starbucks_4, starbucks_5]
starbucks_result = pd.concat(starbucks_ind)


#reset index for the starbucks df 

starbucks_result = starbucks_result.reset_index(drop=True)


#drop any duplicates 

starbucks_result = starbucks_result.drop_duplicates(subset=['ADDRESS']).reset_index(drop=True)
starbucks_result.ADDRESS.nunique()

#add traffic rank into starbucks results 

starbucks_res= pd.merge(starbucks_result,top_five[['STATION_FULL','T_RANK']],left_on='STATION', right_on="STATION_FULL", how='left')


#drop the duplicate column of station name and sort by traffic rank 
starbucks_res = starbucks_res.drop(columns=["STATION_FULL"])
starbucks_res = starbucks_res.sort_values("T_RANK")
starbucks_res.head()

#load percent data for top ten station

top_percent = pd.read_csv("./top_percent_2021_in_2019.csv", header = 0)
top_percent.head()


#convert percent and week column to numeric

top_percent["WEEK"] = pd.to_numeric(top_percent["WEEK"])
top_percent["PERCENT"] = pd.to_numeric(top_percent["PERCENT"])
#select data for last week in dataset and sort by top percentage first 

top_per_30 = top_percent[top_percent["WEEK"]== 30].sort_values("PERCENT", ascending=False)

#choose stations with percentage value above the 2021 average

top_per = top_per_30[top_per_30["PERCENT"] > 45.402703].reset_index(drop=True)
top_per.head()

#assign rank to top stations

top_per["P_RANK"] = top_per.index + 1
top_per.head()

#add percent rank into starbucks results 

starbucks_results= pd.merge(starbucks_res,top_per[['STATION_FULL','P_RANK']],left_on='STATION', right_on="STATION_FULL", how='left')

#drop the duplicate column of station name and fill NA with 0 
starbucks_results = starbucks_results.drop(columns=["STATION_FULL"])
starbucks_results = starbucks_results.fillna(10)
starbucks_results = starbucks_results.astype({"P_RANK": int})

starbucks_results.head()


#finally, sort first by percent rank and then by traffic to get complete list of starbucks to focus on
#ranked first by potential growth and then by traffic at location 


starbucks_ranked = starbucks_results.sort_values(['P_RANK', 'T_RANK'], ascending=[True, True]).reset_index(drop=True)
starbucks_ranked

#saving starbucks data for plotting

starbucks_ranked.to_csv("./starbucks_ranked.csv", index=False)
