==> Code/6- Interactive Visualizations.ipynb <==
import pandas as pd
from bokeh.io import output_notebook, show
from bokeh.plotting import gmap, figure, output_file, show
from bokeh.models import GMapOptions, ColumnDataSource, HoverTool
from bokeh.resources import INLINE, CDN
from bokeh.transform import linear_cmap
from bokeh.palettes import Spectral6
from bokeh.io import export_png
from bokeh.embed import file_html

#load starbucks data

starbucks = pd.read_csv("./starbucks_ranked.csv", )
starbucks.head()

starbucks["ORDER"] = starbucks.index+1
starbucks.head()

#creating a bounding box for plotting 

BBox = ((starbucks.LONG.min(),   starbucks.LONG.max(),      
         starbucks.LAT.min(), starbucks.LAT.max()))
BBox

lat,lon = ( ((BBox[2]+BBox[3])/2), ((BBox[0]+BBox[1])/2))
lat,lon

GOOGLE_API_KEY = 

output_notebook()

output_file("starbucks.html")
map_options = GMapOptions(lat=lat, lng=lon, map_type="roadmap", zoom=13)


source = ColumnDataSource(data={
    'long': starbucks.LONG,
    'lat' : starbucks.LAT,
    'order':starbucks.ORDER,
    'address': starbucks.ADDRESS,
    'name': starbucks.NAME,})



TOOLTIPS = [("Priority", "@order"),
            ("Name", "@name"),
            ("Address", "@address"),
            
        ]

p = gmap(GOOGLE_API_KEY, map_options, title="Starbucks Locations to Open for in-store Orders in Manhattan")
p.add_tools(HoverTool(tooltips=TOOLTIPS))

p.circle(x="long", y="lat", size=15, fill_color="green", fill_alpha=0.8, source=source)



show(p)

html = file_html(p, CDN, "starbucks_locs")
